+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 renamings in sf06_lf_i2c: 
 - init_driver() changed to sf06_lf_init()
 - naming of i2c address defines changed
+- added prefix to enum types in sf06_lf_i2c header (cmd_id_t -> sf06_lf_cmd_id_t, inv_flow_scale_factors -> sf06_lf_inv_flow_scale_factors)
 
 ## [0.1.0] - 2022-03-30